fix(changmen): updateOrderBind 显式 $1::bigint 避免 pg 类型推断冲突
日志显示 55 次 "column link is of type bigint but expression is of type uuid"。
linkVal 超过 int4 范围时 pg 库可能将其序列化为字符串，PostgreSQL
在同一参数列表中与 UUID 类型参数混淆。加 ::bigint 显式转换。

+++ server/db/rds/orders_store.js
@@ -405,7 +405,7 @@ export async function updateOrderBind(orderId, userId, link, opts = {}) {
     if (!prev) return false;
     const updateWhere = offsetSqlPlaceholders(where, 1);
     const res = await pool.query(
-      `UPDATE orders SET link = $1 WHERE ${updateWhere}`,
+      `UPDATE orders SET link = $1::bigint WHERE ${updateWhere}`,
       [linkVal, ...params],
     );
     if (res.rowCount > 0 && shouldFireOrderBoundHook(prev, linkVal) && _ordersBoundHook) {
